+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
 
+- [Patch] Improve logger call on different API versions [#664](https://github.com/Shopify/shopify-api-js/pull/664)
+
 ## [6.1.0] - 2023-01-05
 
 - [Minor] Allow api version overrides [#660](https://github.com/Shopify/shopify-api-js/pull/660)